Move some commands to application
