+++ application/cmd/insolar/add_migration_addresses.go
