+++ application/cmd/insolar/create_default_configs.go
