+++ application/cmd/insolar/free_migration_addresses.go
