+++ application/cmd/insolar/main.go
