+++ functest/0_test.go
@@ -21,7 +21,7 @@ package functest
 import (
 	"testing"
 
-	"github.com/insolar/insolar/testutils/launchnet"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 )
 
 // This test file contains tests what always must be first in the package.
