+++ functest/common_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/insolar/insolar/application/api/requester"
-	"github.com/insolar/insolar/testutils/launchnet"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 )
 
 func TestGetRequest(t *testing.T) {
