+++ functest/deposit_transfer_test.go
@@ -28,8 +28,8 @@ import (
 
 	"github.com/insolar/insolar/application/api"
 	"github.com/insolar/insolar/application/api/requester"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	"github.com/insolar/insolar/testutils/launchnet"
 )
 
 // TODO: https://insolar.atlassian.net/browse/WLT-768
