+++ functest/get_balance_test.go
@@ -24,8 +24,8 @@ import (
 
 	"github.com/insolar/insolar/application/api/requester"
 
+	"github.com/insolar/insolar/application/testutils/launchnet"
 	"github.com/insolar/insolar/insolar/gen"
-	"github.com/insolar/insolar/testutils/launchnet"
 
 	"github.com/stretchr/testify/require"
 )
