+++ functest/get_node_ref_test.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/insolar/insolar/application/api/requester"
 
-	"github.com/insolar/insolar/testutils/launchnet"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 
 	"github.com/stretchr/testify/require"
 )
