+++ functest/get_status_test.go
@@ -21,7 +21,7 @@ package functest
 import (
 	"testing"
 
-	"github.com/insolar/insolar/testutils/launchnet"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 
 	"github.com/stretchr/testify/require"
 )
