+++ functest/helloworld_test.go
@@ -28,9 +28,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/insolar/insolar/application/api/requester"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 	"github.com/insolar/insolar/insolar"
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	"github.com/insolar/insolar/testutils/launchnet"
 )
 
 type HelloWorldInstance struct {
