+++ functest/member_create_test.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/insolar/insolar/application/api/requester"
 
-	"github.com/insolar/insolar/testutils/launchnet"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 
 	"github.com/stretchr/testify/require"
 )
