+++ functest/new_api_test.go
@@ -32,7 +32,7 @@ import (
 	"github.com/insolar/insolar/platformpolicy"
 
 	"github.com/insolar/insolar/application/api/requester"
-	"github.com/insolar/insolar/testutils/launchnet"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 )
 
 func contractError(body []byte) error {
