+++ functest/send_money_test.go
@@ -27,8 +27,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/insolar/insolar/application/api/requester"
+	"github.com/insolar/insolar/application/testutils/launchnet"
 	"github.com/insolar/insolar/insolar/gen"
-	"github.com/insolar/insolar/testutils/launchnet"
 )
 
 const times = 5
